Add src/build_util.zig
This file contains helpers for building a VST.  For example, on macOS a
.vst bundle is automatically generated with a proper Info.plist and
PkgInfo file.

---

Future Outlook

I'm currently using Ableton Live to test the generated VST.  However,
this is huge pain in the ass, because Ableton appears to cache plugins.
This means that I have to restart it for every simple change, which
takes some time and is especially annoying in the early stages of this
library. All other acceptable VST hosts I tried on macOS have the same
issue. Maybe it's related to the .vst CFBundleVersion? Hm...

Anyways: I love the idea of implementing a separate build step that
executes your VST and handles MIDI I/O and Audio Output on all major
platforms. Basically, a simple VST host that could simply be launched by
`zig build run-vst` or whatever. Even though this isn't of great use
when you test something along the lines of DAW-interop, I still think
that it would be a big quality-of-life improvement.
Especially with music it is important to have a quick feedback loop, in
my opinion. You want to be able to immediately listen to changes.